Revert "Bring general sketch commits upstream"
